requirements.txt
bcrypt==4.3.0
httpx==0.28.1
PyJWT==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20
starlette==0.47.3
uvicorn==0.35.0